% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinalSize.R
\name{getFinalSize}
\alias{getFinalSize}
\title{getFinalSize}
\usage{
getFinalSize(
  vacTime,
  vacPortion,
  popSize,
  R0,
  recoveryRate,
  contactRatio,
  contactWithinGroup,
  suscRatio
)
}
\arguments{
\item{vacTime}{time after first case at which all vaccinations are delivered}

\item{vacPortion}{fraction of each population vaccinated}

\item{popSize}{the size of each population}

\item{R0}{overall basic reproduction number}

\item{recoveryRate}{inverse of mean infectious period (same time units as vacTime)}

\item{contactRatio}{ratio of 2nd group's : 1st group's overall contact rate}

\item{contactWithinGroup}{fraction of each group's contacts that are exclusively within group}

\item{suscRatio}{ratio of 2nd group's : 1st group's susceptibility to infection per contact}
}
\description{
This function calculates the final sizes of the outbreaks.
}
\details{
Run this function to run the model.
}
